backend/requirements.txt
fastapi[all]
uvicorn
google-cloud-translate
python-jose
pymongo
motor
python-dotenv
passlib[bcrypt]
pyjwt
openai
python-docx
PyPDF2
reportlab

